+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+- Added extension `Object.isJSAny` and `Object.objectDartify()`;
+
+- CI: test with `dart2js` and `dart2wasm` (on Chrome).
+
 ## 1.0.0
 
 - Initial version.